Namespaces renaming
