+++ Mapper.Playground/Models/ActorMapConfiguration.cs
@@ -1,11 +1,14 @@
-namespace Mapper.Playground.Models;
+using SimpleTools.Mapper.Configurations;
+
+namespace SimpleTools.MapperPlayground.Models;
 
 public class ActorMapConfiguration : MapConfiguration
 {
     protected override void Prepare()
     {
-        SetOptions<Actor, ActorModel>(options => options
-                .Ignore(_ => _.Name)
-        );
+        SetOptions<Actor, ActorModel>(options =>
+        {
+            options.Ignore(_ => _.Name);
+        });
     }
 }
